Automatically build and push trellis-ext-db image
But only 1) on changes to the `master` branch, and 2) when the changed files include the Dockerfile or any files under the `trellis` directory. This gives us automatic image updates without triggering unnecessary builds/pushes (e.g., if the changes are to the sinopia_client or the README, which do not require a new image).